reformatted code, moved classes to their appropriate packages
