+++ src/main/java/application/auth/models/About.java
@@ -1,51 +1,49 @@
 package application.auth.models;
 
 public class About {
-	
-	private String name;
 
-	private String parentRepo;
+    private String name;
 
-	private String description;
-	
-	public About() {
-		
-	}
-	
-	public About(String name, String parentRepo, String description) {
-		this.setName(name);
-		this.setParentRepo(parentRepo);
-		this.setDescription(description);
-	}
+    private String parentRepo;
 
-	public String getName() {
-		return name;
-	}
+    private String description;
 
-	public void setName(String name) {
-		this.name = name;
-	}
+    public About() {
 
-	public String getParentRepo() {
-		return parentRepo;
-	}
+    }
 
-	public void setParentRepo(String parentRepo) {
-		this.parentRepo = parentRepo;
-	}
+    public About(String name, String parentRepo, String description) {
+        this.setName(name);
+        this.setParentRepo(parentRepo);
+        this.setDescription(description);
+    }
 
-	public String getDescription() {
-		return description;
-	}
+    public String getName() {
+        return name;
+    }
 
-	public void setDescription(String description) {
-		this.description = description;
-	}
+    public void setName(String name) {
+        this.name = name;
+    }
 
-	@Override
-	public String toString() {
-		return "About [name=" + name + ", parentRepo=" + parentRepo + ", description=" + description + "]";
-	}
-	
+    public String getParentRepo() {
+        return parentRepo;
+    }
 
+    public void setParentRepo(String parentRepo) {
+        this.parentRepo = parentRepo;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public String toString() {
+        return "About [name=" + name + ", parentRepo=" + parentRepo + ", description=" + description + "]";
+    }
 }
