--- src/main/java/application/auth/models/JwtConfig.java
@@ -1,22 +1,0 @@
-package application.auth.models;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@EnableConfigurationProperties
-@ConfigurationProperties(prefix="jwt")
-public class JwtConfig {
-	
-	private String sharedSecret;
-	
-	public String getSharedSecret() {
-		return sharedSecret;
-	}
-	
-	public void setSharedSecret(String sharedSecret) {
-		this.sharedSecret = sharedSecret;
-	}
-
-}
